feat(shell-tools): F4.6.7 phase 2+3 — sink ShellTool body and delete OsExecutor shim
把 desktop-client/ironclaw/src/tools/builtin/shell.rs 里 ~950 行的 ShellTool
本体、以及 sandbox/os_executor.rs 里那一层 SandboxedExecutor shim，全部
下沉到 dasclaw_shell_tools crate。桌面端只保留：
- shell.rs 15 行的 `pub use` 再导出；
- sandbox/config.rs 里新增的 legacy_policy_to_cap() 翻译函数（3-variant
  → 4-variant dasclaw_workspace_cap policy）；
- app.rs::activate_sandbox 里用翻译函数 + SandboxedShellExecutor 装好
  BootstrapContext。

44 个 ShellTool / helpers / sandboxed_executor 单元测试随类型一起进入
crate，desktop-client 里所有 shell/sandbox/req_sandbox_w3 集成测试
（79 个）继续通过。

ADR-156 §6.3 F4.6 builtin-tool landing
ADR-114 cross-cuts 类 A
ADR-153 headless agent framework

Closes #785
